Куча кода, лень описывать всё

+++ ConThing/CatalogForm.cs
@@ -4,6 +4,7 @@ using System.ComponentModel;
 using System.Data;
 using System.Data.SQLite;
 using System.Drawing;
+using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -19,9 +20,13 @@ namespace ConThing {
 		/// </summary>
 		private SQLiteConnection connection;
 
+		private ItemPreviewForm previewForm;
+
 		public CatalogForm(SQLiteConnection connection) {
 			this.connection = connection;
 			InitializeComponent();
+
+			previewForm = new ItemPreviewForm();
 		}
 
 		/// <summary>
@@ -43,7 +48,8 @@ namespace ConThing {
 					((long)reader[0]).ToString(),
 					(string)reader[1],
 					string.Format("{0:F2} ₽", (double)reader[2]),
-					((int)reader[3]).ToString()
+					((int)reader[3]).ToString(),
+					reader.IsDBNull(4) ? "" : (string)reader[4]
 				}));
 			}
 
@@ -52,6 +58,8 @@ namespace ConThing {
 
 			// пересчитываем ширину столбцов
 			UpdateColumnWidth();
+
+			previewForm.Show();
 		}
 
 		/// <summary>
@@ -64,15 +72,24 @@ namespace ConThing {
 			// если при вызове формы не вернули ok, то уходим
 			if (form.ShowDialog() != DialogResult.OK) return;
 
+			var dir = Directory.CreateDirectory("imgs");
+			var fi = new FileInfo(form.ImagePath);
+			var path = "";
+			if (fi.Exists) {
+				path = dir.FullName + "\\" + RandomName() + fi.Extension;
+				fi.CopyTo(path);
+			}
+
 			// вставим элемент в БД
 
 			// открываем соединение
-			var com = new SQLiteCommand("insert into items(name, price, quantity) values(@Name, @Price, @Quantity); select last_insert_rowid();", connection);
+			var com = new SQLiteCommand("insert into items(name, price, quantity, img_path) values(@Name, @Price, @Quantity, @ImgPath); select last_insert_rowid();", connection);
 
 			// указываем параметры
 			com.Parameters.Add("@Name", DbType.String).Value = form.ItemName;
 			com.Parameters.Add("@Price", DbType.Double).Value = form.Price;
 			com.Parameters.Add("@Quantity", DbType.Int32).Value = form.Quantity;
+			com.Parameters.Add("@ImgPath", DbType.String).Value = path;
 			
 			// получаем id
 			var id = (long)com.ExecuteScalar();
@@ -82,7 +99,8 @@ namespace ConThing {
 				id.ToString(),
 				form.ItemName,
 				string.Format("{0:F2} ₽", form.Price),
-				form.Quantity.ToString()
+				form.Quantity.ToString(),
+				path
 			}));
 
 			// обновляем ширину столбцов
@@ -105,6 +123,9 @@ namespace ConThing {
 		/// Удаляет выбранные строки.
 		/// </summary>
 		private void DeleteSelectedRows() {
+			if (MessageBox.Show("Действительно хотите удалить эти элементы?", "*_*", 
+				MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.No) return;
+
 			for (int i = lstCatalog.SelectedItems.Count - 1; i >= 0; i--) {
 				var row = lstCatalog.SelectedItems[i];
 				DeleteById(long.Parse(row.SubItems[0].Text));
@@ -130,6 +151,52 @@ namespace ConThing {
 		private void UpdateColumnWidth() {
 			for (int i = 0; i < lstCatalog.Columns.Count; i++)
 				lstCatalog.Columns[i].Width = -2;
+		}
+
+		private string RandomName() {
+			var rand = new Random();
+			var res = "";
+			for (int i = 0; i < 16; i++)
+				res += rand.Next(16).ToString("x");
+			return res;
+		}
+
+		private void CatalogForm_FormClosing(object sender, FormClosingEventArgs e) {
+			if (!previewForm.IsDisposed)
+				previewForm.Close();
+		}
+
+		private void lstCatalog_ItemSelectionChanged(object sender, ListViewItemSelectionChangedEventArgs e) {
+			previewForm.SetImagePath(e.Item.SubItems[4].Text);
+		}
+
+		private void lstCatalog_DoubleClick(object sender, EventArgs e) {
+			var row = lstCatalog.SelectedItems[0];
+			var form = new EditCatalogItemForm(row);
+			if (form.ShowDialog() != DialogResult.OK) return;
+
+			var dir = Directory.CreateDirectory("imgs");
+			var fi = new FileInfo(form.ImagePath);
+			var path = "";
+			if (fi.Exists) {
+				path = dir.FullName + "\\" + RandomName() + fi.Extension;
+				fi.CopyTo(path);
+			}
+
+			var com = new SQLiteCommand("update items set name=@Name,price=@Price,quantity=@Quantity,img_path=@ImgPath where id=@Id;", connection);
+
+			com.Parameters.Add("@Name", DbType.String).Value = form.ItemName;
+			com.Parameters.Add("@Price", DbType.Double).Value = form.Price;
+			com.Parameters.Add("@Quantity", DbType.Int32).Value = form.Quantity;
+			com.Parameters.Add("@ImgPath", DbType.String).Value = path;
+			com.Parameters.Add("@Id", DbType.Int64).Value = long.Parse(row.SubItems[0].Text);
+
+			com.ExecuteNonQuery();
+
+			row.SubItems[1].Text = form.ItemName;
+			row.SubItems[2].Text = string.Format("{0:F2} ₽", form.Price);
+			row.SubItems[3].Text = form.Quantity.ToString();
+			row.SubItems[4].Text = path;
 		}
 	}
 }
